chore: clean fish functions
Former-commit-id: df16fdc5208e8b5885c1c8dbe7158fc84dad0927

+++ .config/fish/functions/reml.fish
@@ -1,5 +1,10 @@
 function reml --description 'uninstall large packages need to be upgraded'
-    set -l pkgs (pacman -Qui | grep -E '^(Name|Installed)' | cut -f2 -d':' | paste - - | column -t | grep MiB | sort -nrk 2 | bat)
-    set -l to_remove (echo $pkgs | fzf -m | cut -f1 -d' ')
-    sudo pacman -Rscd $to_remove
+    set -l pkgs (pacman -Qui | awk -F ': ' '
+        /^Name/{name=$2}
+        /^Installed/{size=$2}
+        /MiB$/{print name, size}' | sort -nrk 2 | column -t)
+    set -l to_remove (printf '%s\n' $pkgs | fzf -m | cut -f1 -d' ')
+    if test (count $to_remove) -ne 0
+        sudo pacman -Rscd $to_remove
+    end
 end
